callautomation-playandrecognize/requirements.txt
Flask>=2.3.2
azure-eventgrid==4.11.0
azure-communication-callautomation==1.1.0a20230802001
